ml_eng_test_solution/requirements.txt
uvicorn
fastapi
fastapi[all]
torch
torchvision
pillow
matplotlib==3.0.3
lmdb
opencv-python
scikit-image
svgpathtools
shapely

paddlepaddle
paddleocr